+++ rumqttd/src/server/broker.rs
@@ -1,3 +1,4 @@
+use crate::link::alerts::{self};
 // use crate::link::bridge;
 use crate::link::console::ConsoleLink;
 use crate::link::network::{Network, N};
@@ -11,7 +12,7 @@ use crate::protocol::ws::Ws;
 use crate::protocol::Protocol;
 #[cfg(any(feature = "use-rustls", feature = "use-native-tls"))]
 use crate::server::tls::{self, TLSAcceptor};
-use crate::{meters, ConnectionSettings, GetMeter, Meter};
+use crate::{meters, ConnectionSettings, Filter, GetMeter, Meter};
 use flume::{RecvError, SendError, Sender};
 use std::net::SocketAddr;
 use std::sync::Arc;
@@ -129,6 +130,24 @@ impl Broker {
         Ok(link)
     }
 
+    /// Alerts of different kind are published on topic mentioned below. MQTT wildcards
+    /// can be used to subscribe to multiple types of alerts at the same time. This are seperated
+    /// from normal topics, so they are only available over AlertsLink
+    ///
+    /// /alerts/<alert_type>/<connection_id>
+    /// alert_kind can be of types:
+    ///     - "connect"
+    ///     - "disconnect"
+    ///     - "subscribe"
+    ///     - "unsubscribe"
+    ///
+    /// To get all the types of alerts for all connection, you can subscribe to:
+    /// - /alerts/#
+    pub fn alerts(&self, filters: Vec<Filter>) -> Result<alerts::AlertsLink, alerts::LinkError> {
+        let link = alerts::AlertsLink::new(self.router_tx.clone(), filters)?;
+        Ok(link)
+    }
+
     pub fn link(&self, client_id: &str) -> Result<(LinkTx, LinkRx), local::LinkError> {
         // Register this connection with the router. Router replies with ack which if ok will
         // start the link. Router can sometimes reject the connection (ex max connection limit)
@@ -166,7 +185,7 @@ impl Broker {
                 let runtime = runtime.enable_all().build().unwrap();
 
                 runtime.block_on(async {
-                    if let Err(e) = server.start(false).await {
+                    if let Err(e) = server.start(LinkType::Remote).await {
                         error!(error=?e, "Remote link error");
                     }
                 });
@@ -181,7 +200,7 @@ impl Broker {
                 let runtime = runtime.enable_all().build().unwrap();
 
                 runtime.block_on(async {
-                    if let Err(e) = server.start(false).await {
+                    if let Err(e) = server.start(LinkType::Remote).await {
                         error!(error=?e, "Remote link error");
                     }
                 });
@@ -203,7 +222,7 @@ impl Broker {
                 let runtime = runtime.enable_all().build().unwrap();
 
                 runtime.block_on(async {
-                    if let Err(e) = server.start(true).await {
+                    if let Err(e) = server.start(LinkType::Shadow).await {
                         error!(error=?e, "Remote link error");
                     }
                 });
@@ -270,6 +289,13 @@ impl Broker {
     }
 }
 
+#[derive(Copy, Clone)]
+pub enum LinkType {
+    #[cfg(feature = "websockets")]
+    Shadow,
+    Remote,
+}
+
 struct Server<P> {
     config: ServerSettings,
     router_tx: Sender<(ConnectionId, Event)>,
@@ -303,7 +329,7 @@ impl<P: Protocol + Clone + Send + 'static> Server<P> {
         Ok((Box::new(stream), None))
     }
 
-    async fn start(&self, shadow: bool) -> Result<(), Error> {
+    async fn start(&self, link_type: LinkType) -> Result<(), Error> {
         let listener = TcpListener::bind(&self.config.listen).await?;
         let delay = Duration::from_millis(self.config.next_connection_delay_ms);
         let mut count: usize = 0;
@@ -347,16 +373,16 @@ impl<P: Protocol + Clone + Send + 'static> Server<P> {
             count += 1;
 
             let protocol = self.protocol.clone();
-            match shadow {
+            match link_type {
                 #[cfg(feature = "websockets")]
-                true => task::spawn(shadow_connection(config, router_tx, network).instrument(
-                    tracing::error_span!(
+                LinkType::Shadow => task::spawn(
+                    shadow_connection(config, router_tx, network).instrument(tracing::info_span!(
                         "shadow_connection",
                         client_id = field::Empty,
                         connection_id = field::Empty
-                    ),
-                )),
-                _ => task::spawn(
+                    )),
+                ),
+                LinkType::Remote => task::spawn(
                     remote(config, tenant_id, router_tx, network, protocol).instrument(
                         tracing::error_span!(
                             "remote_link",