+++ CHANGELOG.md
@@ -12,7 +12,9 @@ rumqttd
 - Don't allow client's with empty client_id (#546)
 - Disconnect existing client on a new connection with similar client_id (#546)
 - Skip client certificate verification when using native-tls because it doesn't support it (#550)
+- Add AlertLink to get alerts about router events (#538)
 - Add basic username and password authentication (#553)
+
 
 ### R19
 ---